// src/types/companies.types.ts
export type CompanyRow = {
  id: string;
  name: string;
  phone: string;
  email: string | null;
  category: string | null;
  address: string | null;
  city: string | null;
  state: string | null;
  country: string | null;
  zipcode: string | null;
  messageCredits: number;
  createdAt: string;
  updatedAt: string;
};